update test.py from example.py
